Add authentication
Install and use firebase to create authentication in app. Now user
can sing in, sing up, recovery password, change email, change password
and delete account. Create custom hook - useAuth. There are all functions,
erros handling and more related to authentication. Create SignIn,
SignUp, UserAccount and PasswordRecovery views and styles for them.
UserAuccount compononet checks if currentUser  exists, if so, it displays
the account but if not, it moves to SignIn view and here user can
sign in. If data is correct and account is exists it moves to
UserAccoutn view, but if not, it shows error. SignUp view works
similarly.